# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: CVE Daily Report API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /apple/cves/{platform}:
    get:
      operationId: getAppleCves
      tags: [apple]
      summary: Get per-CVE detail for the latest Apple security release
      description: Returns per-CVE description, NVD link, actively-exploited flag, and version-specific annotation for the latest release, parsed from the ios-security-vulnerability-formatter service's full report.
      parameters:
        - name: platform
          in: path
          required: true
          schema:
            type: string
            enum: [ios, macos]
      responses:
        "200":
          description: Per-CVE detail for the latest release
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppleCvesResult"
  /apple/history/{platform}:
    get:
      operationId: getAppleHistory
      tags: [apple]
      summary: Get historical Apple security releases
      description: Returns every distinct (platform, version) release api-server has ever seen for this platform, newest first — an accumulating changelog independent of the upstream tool, which only ever reports the current latest release(s). Empty if no persistent store is configured.
      parameters:
        - name: platform
          in: path
          required: true
          schema:
            type: string
            enum: [ios, macos]
      responses:
        "200":
          description: Release history, newest first
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppleHistoryResult"
...
components:
  schemas:
...
    AppleCve:
      type: object
      required: [cveId, description, activelyExploited, versionNote, nvdUrl]
      properties:
        cveId:
          type: string
        description:
          type: ["string", "null"]
        activelyExploited:
          type: boolean
        versionNote:
          type: ["string", "null"]
        nvdUrl:
          type: string
    AppleCvesResult:
      type: object
      required: [cves]
      properties:
        cves:
          type: array
          items:
            $ref: "#/components/schemas/AppleCve"
    AppleReleaseHistoryEntry:
      type: object
      required: [platform, version, updateName, releaseDate, cveCount, securityInfoUrl, activelyExploited, firstSeenAt, lastSeenAt]
      properties:
        platform:
          type: string
          enum: [ios, macos]
        version:
          type: string
        updateName:
          type: ["string", "null"]
        releaseDate:
          type: string
        cveCount:
          type: integer
        securityInfoUrl:
          type: ["string", "null"]
        activelyExploited:
          type: boolean
        firstSeenAt:
          type: string
        lastSeenAt:
          type: string
    AppleHistoryResult:
      type: object
      required: [history]
      properties:
        history:
          type: array
          items:
            $ref: "#/components/schemas/AppleReleaseHistoryEntry"
...