# lib/api-zod/src/generated/types/appleHistoryResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CVE Daily Report API
 * OpenAPI spec version: 0.1.0
 */
import type { AppleReleaseHistoryEntry } from './appleReleaseHistoryEntry';

export interface AppleHistoryResult {
  history: AppleReleaseHistoryEntry[];
}