# lib/api-zod/src/generated/types/appleReleaseHistoryEntry.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CVE Daily Report API
 * OpenAPI spec version: 0.1.0
 */
import type { AppleReleaseHistoryEntryPlatform } from './appleReleaseHistoryEntryPlatform';

export interface AppleReleaseHistoryEntry {
  platform: AppleReleaseHistoryEntryPlatform;
  version: string;
  /** @nullable */
  updateName: string | null;
  releaseDate: string;
  cveCount: number;
  /** @nullable */
  securityInfoUrl: string | null;
  activelyExploited: boolean;
  firstSeenAt: string;
  lastSeenAt: string;
}